Fix checkstyle
