+++ build.gradle
@@ -14,6 +14,10 @@ dependencies {
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
 }
 
+checkstyle {
+    toolVersion = '10.2'
+}
+
 test {
     useJUnitPlatform()
