+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm SlayBot.SlayBot
+Hello! I'm slaybot.slaybot
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
